+++ src/components/admin/compose/ComposeSidebarRight.vue
@@ -1,47 +1,16 @@
 <template lang="pug">
   nav.ComposeSidebarRight
     .section.is-small
-      menu-ui(:groups='menuGroups')
-
-    //- .menu.animated.slideInLeft
+      component-edit()
 </template>
 
 <script>
-import { MenuUi } from '@/components/ui'
+import { ComponentEdit } from '@/components/admin/compose/components'
 
 export default {
   name: 'ComposeSidebarRight',
   components: {
-    MenuUi,
-  },
-  data () {
-    return {
-      menuGroups: [
-        {
-          label: 'General',
-          items: [
-            {
-              text: 'Dashboard',
-              icon: 'tachometer',
-              link: {
-                type: 'router',
-                name: 'DashboardAdmin',
-              },
-            },
-            {
-              text: 'Pages',
-              icon: 'files-o',
-              link: {
-                type: 'router',
-                name: 'PagesList',
-              },
-            },
-          ],
-        },
-      ],
-    }
-  },
-  methods: {
+    ComponentEdit,
   },
 }
 </script>
@@ -54,9 +23,4 @@ export default {
 
   .section.is-small
     padding 1.5rem
-
-  .menu-list
-    li
-      a
-        transition background .2s linear
 </style>