+++ src/components/admin/compose/components/index.js
@@ -1,0 +1,7 @@
+import ComponentEdit from '@/components/admin/compose/components/ComponentEdit'
+import ComponentListAdd from '@/components/admin/compose/components/ComponentListAdd'
+
+export {
+  ComponentEdit,
+  ComponentListAdd
+}